+++ cpsc-471-project/Controllers/ProjectsController.cs
@@ -2,25 +2,24 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
+
+using cpsc_471_project.Authentication;
 using cpsc_471_project.Models;
-using Microsoft.EntityFrameworkCore.ChangeTracking.Internal;
+
 
 namespace cpsc_471_project.Controllers
 {
     [Route("api/resumes")]
     [ApiController]
-    public class ProjectsController : ControllerBase
+    public class ProjectsController : ResumeSectionController
     {
-        private readonly JobHunterDBContext _context;
+        public ProjectsController(JobHunterDBContext context, UserManager<User> userManager) : base(context, userManager) { }
 
-        public ProjectsController(JobHunterDBContext context)
-        {
-            _context = context;
-        }
-
+        [Authorize]
         [HttpPatch("{resumeId}/projects/{order}")]
         public async Task<IActionResult> PatchProject(long resumeId, long order, ProjectDTO projectDTO)
         {
@@ -33,6 +32,11 @@ namespace cpsc_471_project.Controllers
             if (order != sanitizedProject.Order)
             {
                 return BadRequest("subsection order in query params does not match subsection order in body");
+            }
+
+            if (!await ResumeAccessAuthorized(sanitizedProject.ResumeId))
+            {
+                return GenerateResumeNotFoundError(sanitizedProject.ResumeId);
             }
 
             if (!ProjectExists(resumeId, order))
@@ -53,24 +57,35 @@ namespace cpsc_471_project.Controllers
             return AcceptedAtAction("PatchProject", new { resumeId = sanitizedProject.ResumeId, order = sanitizedProject.Order }, projectDTO);
         }
 
-        [HttpPost("{reusmeId}/Projects")]
+        [HttpPost("{resumeId}/Projects")]
         public async Task<ActionResult<ProjectDTO>> PostProject(ProjectDTO projectDTO)
         {
             Project sanitizedProject = DTOToProject(projectDTO);
+
+            if (!await ResumeAccessAuthorized(sanitizedProject.ResumeId))
+            {
+                return GenerateResumeNotFoundError(sanitizedProject.ResumeId);
+            }
+
             _context.Projects.Add(sanitizedProject);
             await _context.SaveChangesAsync();
 
-            return CreatedAtAction("PostProject", new { reusmeId = sanitizedProject.ResumeId, order = sanitizedProject.Order }, projectDTO);
+            return CreatedAtAction("PostProject", new { resumeId = sanitizedProject.ResumeId, order = sanitizedProject.Order }, projectDTO);
         }
 
         // DELETE: api/Project/5
-        [HttpDelete("{reusmeId}/Projects/{order}")]
-        public async Task<ActionResult<ProjectDTO>> DeleteProject(long reusmeId, long order)
+        [HttpDelete("{resumeId}/Projects/{order}")]
+        public async Task<ActionResult<ProjectDTO>> DeleteProject(long resumeId, long order)
         {
-            var project = await _context.Projects.FindAsync(reusmeId, order);
+            if (!await ResumeAccessAuthorized(resumeId))
+            {
+                return GenerateResumeNotFoundError(resumeId);
+            }
+
+            var project = await _context.Projects.FindAsync(resumeId, order);
             if (project == null)
             {
-                return NotFound();
+                return NotFound("Subsection not found");
             }
 
             _context.Projects.Remove(project);
@@ -79,9 +94,9 @@ namespace cpsc_471_project.Controllers
             return ProjectToDTO(project);
         }
 
-        private bool ProjectExists(long reusmeId, long order)
+        private bool ProjectExists(long resumeId, long order)
         {
-            return _context.Projects.Any(e => (e.ResumeId == reusmeId) && (e.Order == order));
+            return _context.Projects.Any(e => (e.ResumeId == resumeId) && (e.Order == order));
         }
 
         public static ProjectDTO ProjectToDTO(Project project)
